operator: refactor flux controller setup
This commit extracts the setup of flux's controllers from run.go into their own
package. As we have no need for managing each controller individually and they
share many dependencies, creation of them is exposed via single function which
returns a slice of controllers.

This extract will facilitate future test suites which require flux and
eventually the metaphorical sawing off of flux.
